delete simple router implement ，will user tree router
