+++ core/router/error.go
@@ -1,4 +1,4 @@
-package tree_router
+package router
 
 type NotMatch struct {
 	errorStr string
