--- core/router/result.go
@@ -1,8 +1,0 @@
-package router
-
-// 返回值类型
-type MatchResult struct {
-	ParamMap   map[string]string
-	Rest       RestHandler
-	ParamNames []string
-}
