+++ core/router/tree_benchmark_test.go
@@ -1,9 +1,8 @@
 package router
 
 import (
+	"log"
 	"testing"
-
-	"github.com/fudali113/doob/core/register"
 )
 
 var (
@@ -56,8 +55,22 @@ var (
 	}
 )
 
+type testType struct {
+	name string
+	num  int
+}
+
+func (t testType) String() string {
+	return t.name + "----"
+}
+
 func Benchmark_test(b *testing.B) {
-	simpleRouter := &SimpleRouter{}
+	node := &node{
+		class:    normal,
+		value:    nil,
+		handler:  nil,
+		children: make([]*node, 0),
+	}
 	testVar := &RegisterHandler{
 		Handler: &testType{
 			name: "ooo",
@@ -65,23 +78,26 @@ func Benchmark_test(b *testing.B) {
 		},
 	}
 	for _, url := range testData {
-		simpleRouter.Add(url, &SimpleRestHandler{
-			methodHandlerMap: map[string]register.RegisterHandlerType{
-				"get": testVar,
-			},
-		})
+		node.insertChild(url, GetSimpleRestHandler("get", testVar))
 	}
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			oo, _ := simpleRouter.Get("/api/report/index" /**fmt.Sprintf("/aaa/%s/oo", "dd")*/).Rest.GetHandler("get").GetHandler().(*testType)
-			if oo == nil {
+			res := must(node.getRT("/api/report/index", nil)).GetHandler("get").GetHandler()
+			if res == nil {
 				b.Error("path variable method have bug")
 			}
 
-			oo1, _ := simpleRouter.Get("/api/user/barcodes/111-1121-8406/bind_share").Rest.GetHandler("get").GetHandler().(*testType)
-			if oo1 == nil {
+			res1 := must(node.getRT("/api/user/barcodes/111-1121-8406/bind_share", nil)).GetHandler("get").GetHandler()
+			if res1 == nil {
 				b.Error("path variable method have bug")
 			}
 		}
 	})
+}
+
+func must(i RestHandler, e error) RestHandler {
+	if e != nil {
+		log.Panic(e)
+	}
+	return i
 }
