+++ core/router/tree_test.go
@@ -1,33 +1,32 @@
-package tree_router
+package router
 
 import (
 	"log"
 	"testing"
-
-	"github.com/fudali113/doob/core/router"
 )
 
-func TestNode_insertChildren(t *testing.T) {
-
-	node := &node{
+var (
+	testNode = &node{
 		class:    normal,
 		value:    nil,
 		handler:  nil,
 		children: make([]*node, 0),
 	}
+)
 
-	testUrl := "/oooo/lllll/ddddd"
-	node.insertChild(testUrl, &router.SimpleRestHandler{})
-	res, err := node.getRT(testUrl, nil)
-	log.Print(res)
+func TestNode_insertChildren(t *testing.T) {
+
+	testUrl := ""
+	testNode.insertChild(testUrl, &SimpleRestHandler{})
+	res, err := testNode.getRT(testUrl, nil)
 	if err != nil || res == nil {
 		t.Error("Node_insertChildren have bug")
 	}
 
 	testUrl1 := "/oooo/bbbb/**"
 	_testUrl1 := "/oooo/bbbb/o"
-	node.insertChild(testUrl1, &router.SimpleRestHandler{})
-	_, err1 := node.getRT(_testUrl1, nil)
+	testNode.insertChild(testUrl1, &SimpleRestHandler{})
+	_, err1 := testNode.getRT(_testUrl1, nil)
 	if err1 != nil {
 		t.Error("Node_insertChildren have bug 1")
 	}
@@ -35,15 +34,15 @@ func TestNode_insertChildren(t *testing.T) {
 	testUrl2 := "hhhh/{mmm:\\d{3}}/ddddd"
 	_testUrl2 := "hhhh/124/ddddd"
 	paramMap := map[string]string{}
-	node.insertChild(testUrl2, &router.SimpleRestHandler{})
-	_, err2 := node.getRT(_testUrl2, paramMap)
+	testNode.insertChild(testUrl2, &SimpleRestHandler{})
+	_, err2 := testNode.getRT(_testUrl2, paramMap)
 	if err2 != nil {
 		t.Error("Node_insertChildren have bug 3")
 	}
 	if paramMap["mmm"] != "124" {
 		t.Error("Node_insertChildren have bug 3")
 	}
-
+	log.Print(testNode)
 }
 
 func TestNode_Sort(t *testing.T) {
@@ -79,4 +78,9 @@ func TestNode_Sort(t *testing.T) {
 
 	nodeRoot.children = nodes(nodeSlice)
 	nodeRoot.Sort()
+}
+
+func TestNode_getNode(t *testing.T) {
+	getNode := testNode.getNode("oooo")
+	log.Print("-------", getNode)
 }
